pedido implementado, com todos os testes passando

+++ src/app/models/Personalize.js
@@ -5,7 +5,7 @@ class Personalize extends Model {
     super.init(
       {
         name: DataTypes.STRING,
-        price: DataTypes.DECIMAL,
+        price: DataTypes.FLOAT,
       },
       {
         sequelize,
